Monster Jack
Monster Jack is an iOS application game that allows players to hunt for
monsters in an augmented reality, capture and store them in their
libraries for future use. This hunting game can both be single-player
mode or multi-player mode, which the latter is how we can burst the
mobile bubble.